:palm_tree: opening scene laid out
Just need art assets to be able to finish tweaking

+++ Assets/_Scripts/Components/CameraFollow.cs
@@ -11,7 +11,7 @@ namespace Somnium
 
         [Tooltip("How far above the player the camera will be")]
         [SerializeField]
-        private int verticalOffset;
+        private float verticalOffset;
 
         /// <summary>
         /// The target GameObject to follow.
